updated swagger documentation

+++ serverAPI/serverAPI/Controllers/ProductsController.cs
@@ -20,6 +20,10 @@ namespace serverAPI.Controllers
             _db = db;
         }
 
+        /// <summary>
+        /// Return products of the shop.
+        /// </summary>
+        /// <response code="200">Return list of products</response>
         [HttpGet("{shopId}")]
         public JsonResult GetProducts(uint shopId)
         {
@@ -28,6 +32,28 @@ namespace serverAPI.Controllers
             return new JsonResult(Ok(products));
         }
 
+        /// <summary>
+        /// Add new product in shop(for shop's admin).
+        /// </summary>
+        /// <remarks>
+        /// Sample request:
+        ///
+        ///     POST /api/Products
+        ///     {
+        ///         "password":"1",
+        ///         "product": {
+        ///             "shopId":1,
+        ///             "Name":"water", 
+        ///             "Category":"food", 
+        ///             "Cost":10,
+        ///             "Count":10
+        ///         }
+        ///     }
+        ///
+        /// </remarks>
+        /// <response code="200">Return new product</response>
+        /// <response code="400">If request is invalid</response>
+        /// <response code="403">If password incorrect</response>
         [HttpPost]
         public JsonResult PostProduct(ShopAdminProductRequest request)
         {
@@ -43,6 +69,29 @@ namespace serverAPI.Controllers
             return new JsonResult(Ok(product));
         }
 
+        /// <summary>
+        /// Update information about product in shop(for shop's admin).
+        /// </summary>
+        /// <remarks>
+        /// Sample request:
+        ///
+        ///     PUT /api/Products
+        ///     {
+        ///         "password":"1",
+        ///         "product": {
+        ///             "id":1,
+        ///             "shopId":1,
+        ///             "Name":"water", 
+        ///             "Category":"food", 
+        ///             "Cost":10,
+        ///             "Count":10
+        ///         }
+        ///     }
+        ///
+        /// </remarks>
+        /// <response code="200">Return updated product</response>
+        /// <response code="400">If request is invalid</response>
+        /// <response code="403">If password incorrect</response>
         [HttpPut]
         public JsonResult UpdateProduct(ShopAdminProductRequest request)
         {
@@ -61,6 +110,26 @@ namespace serverAPI.Controllers
             return new JsonResult(Ok(product));
         }
 
+        /// <summary>
+        /// Delete product in shop(for shop's admin).
+        /// </summary>
+        /// <remarks>
+        /// Sample request:
+        ///
+        ///     DELETE /api/Products
+        ///     {
+        ///         "password":"1",
+        ///         "product": {
+        ///             "id":1,
+        ///             "shopId":1
+        ///         }
+        ///     }
+        ///
+        /// </remarks>
+        /// <response code="200">Return deleted product</response>
+        /// <response code="400">If request is invalid</response>
+        /// <response code="403">If password incorrect</response>
+        /// <response code="404">If product doesen't exist</response>
         [HttpDelete]
         public JsonResult DeleteProduct(ShopAdminProductRequest request)
         {
@@ -74,11 +143,15 @@ namespace serverAPI.Controllers
             if (product == null)
                 return new JsonResult(NotFound());
 
+            if (product.ShopId != request.Product.ShopId)
+                return new JsonResult(StatusCode(403));
+
             _db.DeleteProduct(product.Id);
 
             return new JsonResult(Ok(product));
         }
 
+        [ApiExplorerSettings(IgnoreApi = true)]
         private bool isRequestValid(ShopAdminProductRequest request, bool DeleteFlag = false)
         {
             if (request == null ||
